+++ lib/Halcyon/Core/Tacky.hs
@@ -5,18 +5,13 @@ import Data.Text ( Text )
 data Program = Program Function
   deriving (Eq, Show)
 
-data Function = Function
-  { name :: Text
-  , body :: [Instruction]
-  } deriving (Eq, Show)
+data Function = Function Text [Instruction]
+  deriving (Eq, Show)
 
 data Instruction
-  = Return { value :: Val }
-  | Unary 
-    { operator :: UnaryOp
-    , src :: Val
-    , dst :: Val 
-    }
+  = Return Val
+  | Unary UnaryOp Val Val
+  | Binary BinaryOp Val Val Val
   deriving (Eq, Show)
 
 data Val
@@ -29,8 +24,22 @@ data UnaryOp
   | Negate
   deriving (Eq, Show)
 
---program = Program(function_definition)
---function_definition = Function(identifier, instruction* body)
---instruction = Return(val) | Unary(unary_operator, val src, val dst)
---val = Constant(int) | Var(identifier)
---unary_operator = Complement | Negate
+data BinaryOp
+  = Add
+  | Subtract
+  | Multiply
+  | Divide
+  | Remainder
+  deriving (Eq, Show)
+
+{-
+program = Program(function_definition)
+function_definition = Function(identifier, instruction* body)
+instruction 
+  = Return(val)
+  | Unary(unary_operator, val src, val dst)
+  | Binary(binary_operator, val src1, val src2, val dst)
+val = Constant(int) | Var(identifier)
+unary_operator = Complement | Negate
+binary_operator = Add | Subtract | Multiply | Divide | Remainder
+-}